% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bfs_codebook.R
\name{get_bfs_codebook}
\alias{get_bfs_codebook}
\title{Get the cori.data.bfs variable codebook}
\usage{
get_bfs_codebook()
}
\value{
A data frame with columns: \code{variable}, \code{label}, \code{unit}, \code{notes}.
}
\description{
Returns documentation for all variables produced by the CORI BFS processing
pipeline.
}
\examples{
get_bfs_codebook()

}
\seealso{
\code{\link[=read_bfs_from_s3]{read_bfs_from_s3()}}
}
