% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bfs_from_s3.R
\name{read_bfs_from_s3}
\alias{read_bfs_from_s3}
\title{Read processed BFS data from S3}
\usage{
read_bfs_from_s3(
  vintage = "latest",
  variables = NULL,
  years = NULL,
  geoids = NULL,
  s3_bucket = "cori.data.bfs",
  s3_path_prefix = ""
)
}
\arguments{
\item{vintage}{Passed to \code{\link[=get_business_applications]{get_business_applications()}}.}

\item{variables}{Passed to \code{\link[=get_business_applications]{get_business_applications()}}.}

\item{years}{Passed to \code{\link[=get_business_applications]{get_business_applications()}}.}

\item{geoids}{Passed to \code{\link[=get_business_applications]{get_business_applications()}}.}

\item{s3_bucket}{Ignored. No longer configurable in the public API.}

\item{s3_path_prefix}{Ignored. No longer configurable in the public API.}
}
\value{
A data frame. See \code{\link[=get_business_applications]{get_business_applications()}} for details.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{read_bfs_from_s3()} is deprecated. Use \code{\link[=get_business_applications]{get_business_applications()}} instead.
}
\seealso{
\code{\link[=get_business_applications]{get_business_applications()}}
}
